+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.4.12 (Unreleased)
 - [Enhancement] Validate presence of `bootstrap.servers` to avoid incomplete partial reconfiguration.
+- [Enhancement] Support `ActiveJob#enqueue_at` via Scheduled Messages feature (Pro).
 
 ## 2.4.11 (2024-09-04)
 - [Enhancement] Validate envelope target topic type for Scheduled Messages.